Move http-get-query-util out of test dir, oops.
